pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "soc_analyst_env"
version = "1.0.0"
description = "Cybersecurity SOC Analyst Environment for OpenEnv — Train AI agents to triage security alerts, investigate incidents, and execute incident response."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SOC Analyst Env Team"}
]

dependencies = [
    "openenv-core>=0.2.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "requests>=2.28.0",
    "openai>=1.0.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.24",
]

[tool.setuptools.packages.find]
include = ["soc_analyst_env*"]
